fix(features): update data type for random embeddings and refine return type in preprocessing
- Changed the data type of random embeddings in `EEGPTFeatureExtractor` from `np.float32` to `np.float64` for improved precision during testing.
- Updated the return type of the `preprocess` method in `FlexibleEEGPreprocessor` to consistently use `MNERaw`, enhancing type safety and clarity in the preprocessing module.
- These changes contribute to better type consistency and align with ongoing efforts to improve code quality across the project.

+++ src/brain_go_brrr/core/features/extractor.py
@@ -84,7 +84,7 @@ class EEGPTFeatureExtractor:
         if self.model is None:
             # Return random embeddings for testing
             logger.warning("Using random embeddings (model not loaded)")
-            embeddings = np.random.randn(len(windows), 512).astype(np.float32)
+            embeddings = np.random.randn(len(windows), 512).astype(np.float64)
         else:
             # Run EEGPT inference
             embeddings = self._run_inference(windows, preprocessed.ch_names)
